fix(race-age): keep heroAge as float in ComputeBaseChance (Codex HIGH)
Codex independent review of the Phase 9b autonomous-run caught a HIGH
regression I introduced in commit 57e9d9b (#179 ComputeBaseChance
extraction). The extraction commit passed (int)hero.Age to the new
helper, truncating fractional age toward zero.

Vanilla DefaultPregnancyModel uses Hero.Age (float) directly in the
ageFactor formula. A 44.9-year-old hero would compute identically to a
44-year-old, materially shifting late-window pregnancy chance. This
was the exact byte-for-byte equivalence the extraction commit's RCA
gate was supposed to enforce.

Fix:
- TaomPregnancyModel.cs: heroAge: (int)hero.Age -> heroAge: hero.Age
- ComputeBaseChance signature: int heroAge -> float heroAge
- Int literals in existing tests implicit-convert to float — no test
  breakage

Added regression test
ComputeBaseChance_FractionalAge_PreservesPrecision: ages 44 / 44.5 / 45
must produce three monotonically-decreasing distinct values. If
heroAge were int-truncated, 44 and 44.5 would match — test would go
red.

RCA at docs/reviews/rca-phase9b-autonomous-codex-review-2026-05-13.md
documents the extraction-without-type-preservation root-cause pattern
and walks through why all 5 deep-review agents missed it (Codex's
independent re-read was the load-bearing safety net). Two new
feedback memories proposed:
- feedback_extraction_must_preserve_types.md
- feedback_test_extracted_helper_with_source_types.md

One MEDIUM Codex finding deferred: Patch35 team filter is
necessary-but-not-sufficient for player-team AI-controlled formations.
Audit #149 specified the team filter as "the simpler fix"; full
threading hardening is Phase 10 candidate, not Phase 9. Tracked in
the RCA.

1957 -> 1958 tests, all passing.

Research: TaleWorlds.CampaignSystem Hero.Age (v1.3.15 verified float)

+++ Main/Features/RaceAge/Models/TaomPregnancyModel.cs
@@ -34,8 +34,14 @@ public class TaomPregnancyModel : DefaultPregnancyModel
         bool playerOrSpouseInvolved = (hero == Hero.MainHero) || (hero.Spouse == Hero.MainHero);
         float fertilityModifier = _raceAgeService.GetFertilityModifier(race);
 
+        // Phase 9b Codex review (2026-05-13): heroAge MUST be float (not int). Vanilla
+        // DefaultPregnancyModel.GetDailyChanceOfPregnancyForHero uses Hero.Age (float) directly
+        // in the ageFactor formula. Truncating to int rounds fractional age toward zero, so a
+        // 44.9-year-old hero would compute as 44 — materially shifting late-window pregnancy
+        // chance. Pre-extract code used `hero.Age` directly; passing (int) here was a regression
+        // in the extraction commit. Keep heroAge as float to preserve byte-for-byte equivalence.
         float baseChance = ComputeBaseChance(
-            heroAge: (int)hero.Age,
+            heroAge: hero.Age,
             comesOfAge: comesOfAge,
             fertilityEnd: fertilityEnd,
             childCount: hero.Children.Count,
@@ -62,7 +68,7 @@ public class TaomPregnancyModel : DefaultPregnancyModel
     /// adapter and move this into IRaceAgeService) is tracked separately as #131.
     /// </summary>
     public static float ComputeBaseChance(
-        int heroAge,
+        float heroAge,
         int comesOfAge,
         int fertilityEnd,
         int childCount,
